Do not assume sizeof(size_t)
This fixes an assumption that sizeof(size_t) == sizeof(uint64_t), which
is not guaranteed. At least on NetBSD/i386, size_t is 32 bit long.

Caught by tests/basics/file-snapshot.t

This is a backport of Ib7620a2ffe8758521886af37bc280101a040d860

BUG: 1138897
Reviewed-on: http://review.gluster.org/8631
Tested-by: Gluster Build System <jenkins@build.gluster.com>

+++ xlators/features/qemu-block/src/qemu-block.h
@@ -40,7 +40,7 @@
 
 typedef struct qb_inode {
 	char fmt[QB_XATTR_VAL_MAX]; /* this is only the format, not "format:size" */
-	size_t size; /* virtual size in bytes */
+	uint64_t size; /* virtual size in bytes */
 	BlockDriverState *bs;
 	int refcnt;
 	uuid_t backing_gfid;
